+++ src/Data/Managers/index.ts
@@ -40,7 +40,6 @@ export * from "./Deploy";
 export * from "./Repair";
 export * from "./GetResourceFacts";
 export * from "./GetAgents";
-export * from "./GetAgentProcess";
 export * from "./GetDesiredStates";
 export * from "./PromoteVersion";
 export * from "./GetVersionResources";
