+++ src/Slices/Agent/Core/Mock.ts
@@ -1,4 +1,4 @@
-import { AgentProcess } from "@/Core";
+import { AgentProcess } from "./Model";
 
 export const data: AgentProcess = {
   sid: "b2c8937e-5359-11ec-9c3f-98e743b19755",
