+++ src/Slices/Agent/Core/Model.ts
