+++ src/Test/Data/index.ts
@@ -22,7 +22,6 @@ export * as Environment from "./Environment";
 export * as ServiceInventory from "./ServiceInventory";
 export * as Facts from "./Facts";
 export * as Agents from "./Agents";
-export * as AgentProcessData from "./AgentProcess";
 export * as DesiredStateVersions from "./DesiredStateVersions";
 export * as Parameters from "./Parameters";
 export * as DesiredStateDiff from "./DesiredStateDiff";
